requirements.txt
streamlit==1.24.1
requests>=2.28
python-dotenv>=1.0.0
wordcloud>=1.8.2.2
matplotlib>=3.6.2
networkx>=2.8.8
python-docx>=0.8.11